fix types

+++ Marlin/src/gcode/cellink/G514.cpp
@@ -14,7 +14,7 @@
  */
 void GcodeSuite::G514()
 {
-    static constexpr float UL_STEPS_FACTOR = (1.0 / (4.6 * 4.6 * 3.14159265)) * ((float[])DEFAULT_AXIS_STEPS_PER_UNIT)[3];
+    static constexpr float UL_STEPS_FACTOR = (1.0 / (4.6 * 4.6 * 3.14159265)) * ((double[])DEFAULT_AXIS_STEPS_PER_UNIT)[3];
     const auto tool = get_target_extruder_from_command();
     const int32_t steps = static_cast<int32_t>(parser.longval('E') * UL_STEPS_FACTOR);
     if (parser.seen('F')) ph_controller.set_extrusion_speed(static_cast<printhead::Index>(tool), parser.value_feedrate());
